MyEvent-ben a debug info kiírás warning mentesítve Néhány elírás javítva

+++ MyDrivers/MyEvent.c
@@ -23,16 +23,16 @@ void MyEvent_printDebugInfo(void)
 {
     Evsys* hw=EVSYS;
     printf("============EVENT SYS INFO==============\n");
-    printf("    CTRLA: %08x\n", hw->CTRLA.reg);
-    printf("  INTPEND: %08x\n", hw->INTPEND.reg);
-    printf("INTSTATUS: %08x\n", hw->INTSTATUS.reg);
-    printf("    SWEVT: %08x\n", hw->SWEVT.reg);
+    printf("    CTRLA: %08x\n", (int)hw->CTRLA.reg);
+    printf("  INTPEND: %08x\n", (int)hw->INTPEND.reg);
+    printf("INTSTATUS: %08x\n", (int)hw->INTSTATUS.reg);
+    printf("    SWEVT: %08x\n", (int)hw->SWEVT.reg);
     printf("\n");
     for(int i=0; i<12; i++)
     {
-       printf("  CHANNEL%02d: %08x -->",i, hw->Channel[i].CHANNEL.reg);
-       printf("  EVGEN=%002d",   hw->Channel[i].CHANNEL.bit.EVGEN);
-       printf("  EDGESEL=%02x", hw->Channel[i].CHANNEL.bit.EDGSEL);
+       printf("  CHANNEL%02d: %08x -->",i, (int)hw->Channel[i].CHANNEL.reg);
+       printf("  EVGEN=%02d",   (int)hw->Channel[i].CHANNEL.bit.EVGEN);
+       printf("  EDGESEL=%02x", (int)hw->Channel[i].CHANNEL.bit.EDGSEL);
 
        switch(hw->Channel[i].CHANNEL.bit.PATH)
        {
@@ -52,7 +52,7 @@ void MyEvent_printDebugInfo(void)
         printf("  %02d-%02d:   ",m, m+9);
         for(int j=0; j<10; j++)
         {
-            printf("%02x ", EVSYS->USER[m].reg);
+            printf("%02x ", (int)EVSYS->USER[m].reg);
             m++;
         }
         printf("\n");
